fix(ensembles): match int64 dtype on recomputed ISOWeek
.dt.isocalendar().week returns UInt32; the rest of the pipeline uses int64
(see train_and_predict.py:142). Add .astype(int) so the ensemble output's
ISOWeek dtype matches per-model outputs and downstream code that reads
predictions.csv doesn't hit dtype-mismatch surprises.

+++ pipelines/dengue/lib/ensembles.py
@@ -90,8 +90,13 @@ class MeanEnsemble:
             )
 
         if "recordDate" in ensembled.columns:
+            # .astype(int) — match the int64 dtype used elsewhere in the pipeline
+            # (train_and_predict.py:142), rather than the UInt32 that
+            # dt.isocalendar().week returns.
             ensembled["ISOWeek"] = (
-                pd.to_datetime(ensembled["recordDate"]).dt.isocalendar().week
+                pd.to_datetime(ensembled["recordDate"])
+                .dt.isocalendar()
+                .week.astype(int)
             )
 
         ensembled["model"] = "ensembleModel"
